[rename/#6] item 패키지 이름 수정
