+++ core/designsystem/src/main/java/com/napzak/market/designsystem/component/productItem/NapzakProductItem.kt
@@ -1,4 +1,4 @@
-package com.napzak.market.designsystem.component.item
+package com.napzak.market.designsystem.component.productItem
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
